feat(java): make modifications to source code addressing comments
