+++ flipt-client-java/src/main/java/io/flipt/client/FliptEvaluationClient.java
@@ -1,5 +1,6 @@
 package io.flipt.client;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -8,6 +9,7 @@ import com.sun.jna.Library;
 import com.sun.jna.Native;
 import com.sun.jna.Pointer;
 import io.flipt.client.models.*;
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Map;
 import java.util.Optional;
@@ -58,6 +60,7 @@ public class FliptEvaluationClient {
     private String url = "http://localhost:8080";
     private AuthenticationStrategy authentication;
     private String reference;
+    private Duration updateInterval;
 
     public FliptEvaluationClientBuilder() {}
 
@@ -76,6 +79,11 @@ public class FliptEvaluationClient {
       return this;
     }
 
+    public FliptEvaluationClientBuilder updateInterval(Duration updateInterval) {
+      this.updateInterval = updateInterval;
+      return this;
+    }
+
     public FliptEvaluationClientBuilder reference(String reference) {
       this.reference = reference;
       return this;
@@ -86,9 +94,47 @@ public class FliptEvaluationClient {
           namespace,
           new EngineOpts(
               Optional.of(url),
-              Optional.of(2),
+              Optional.ofNullable(updateInterval),
               Optional.ofNullable(authentication),
               Optional.ofNullable(reference)));
+    }
+  }
+
+  private static class EvalRequest {
+    private final String namespaceKey;
+
+    private final String flagKey;
+
+    private final String entityId;
+
+    private final Map<String, String> context;
+
+    public EvalRequest(
+        String namespaceKey, String flagKey, String entityId, Map<String, String> context) {
+      this.namespaceKey = namespaceKey;
+      this.flagKey = flagKey;
+      this.entityId = entityId;
+      this.context = context;
+    }
+
+    @JsonProperty("namespace_key")
+    public String getNamespaceKey() {
+      return namespaceKey;
+    }
+
+    @JsonProperty("flag_key")
+    public String getFlagKey() {
+      return flagKey;
+    }
+
+    @JsonProperty("entity_id")
+    public String getEntityId() {
+      return entityId;
+    }
+
+    @JsonProperty("context")
+    public Map<String, String> getContext() {
+      return context;
     }
   }
